backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
pdfplumber==0.11.0
python-multipart==0.0.9
pytest==8.2.0
httpx==0.27.0
python-dotenv>=1.0.0
